requirements.txt
flask==0.12.2
requests==2.18.4
paramiko==2.4.0
mock==2.0.0
PyYAML==3.12
psycopg2==2.7.5